refactor: extract budgetCache utils
